+++ src/main/java/Alden/AddDeadlineCommand.java
@@ -1,3 +1,5 @@
+package Alden;
+
 public class AddDeadlineCommand extends Command {
     private String userInput;
