+++ src/main/java/Alden/AldenException.java
@@ -1,3 +1,4 @@
+package Alden;
 public class AldenException extends Exception {
     public AldenException(String message) {
         super(message);
