+++ src/main/java/Alden/Command.java
@@ -1,3 +1,4 @@
+package Alden;
 public abstract class Command {
     public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws AldenException;
 }
