+++ src/main/java/Alden/Event.java
@@ -1,3 +1,5 @@
+package Alden;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
