+++ src/main/java/Alden/ExitCommand.java
@@ -1,3 +1,5 @@
+package Alden;
+
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
