+++ src/main/java/Alden/Ui.java
@@ -1,3 +1,4 @@
+package Alden;
 public class Ui {
     public void showWelcome() {
         System.out.println("____________________________________________________________");
